requirements.txt
# Core runtime
PyYAML>=6.0.2
numpy>=1.26.4
sounddevice>=0.4.7
soundcard>=0.4.3
webrtcvad-wheels>=2.0.14

# Speech-to-Text
faster-whisper>=1.1.1

# Translation
transformers>=4.49.0
sentencepiece>=0.2.0
torch>=2.3.0
openai>=1.69.0

# Text-to-Speech
pyttsx3>=2.98

# Optional runtime extensions
websockets>=12.0

# Optional alternative TTS provider
# TTS>=0.22.0
